refactor: 도메인 패키지 경로 분리
도메인 클래스 도메인 패키지 경로로 이동
