+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,15 +1,10 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static lotto.LottoResult.PRIZE_MONEY;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoResultTest {
